+++ api/irfaran/transfer.py
@@ -47,7 +47,6 @@ PORTABLE_SETTINGS = frozenset(
         "search_tracks",
         "search_coordinates",
         "search_plus_codes",
-        "search_plus_codes_short",
     }
 )